refactor(hotels): split the room description extractors out
booking_rooms.go was over the project's 800-line limit. The extractors
already sat behind their own section comment, so they move to
booking_rooms_text.go unchanged: no renames, no exported identifiers,
no behaviour change. 832 lines becomes 605 plus 234.
